Goods in fixes (#367)
* default to one label

* allow decimal qties to be booked in... I think

* limit label qties to 2dp

* show qc warning if part is in qc

* show qc warning

* actually allow decimal qty bookins

* fix test

+++ src/Domain.LinnApps/GoodsIn/GoodsInService.cs
@@ -73,7 +73,7 @@
             int createdBy,
             string partNumber,
             string manufacturersPartNumber,
-            int qty,
+            decimal qty,
             int? orderNumber,
             int? orderLine,
             int? loanNumber,
@@ -392,7 +392,7 @@
             int numberOfLines,
             string qcState,
             int reqNumber,
-            string kardexLocation,
+            string kardexLocation,  
             IEnumerable<GoodsInLabelLine> lines)
         {
             var message = string.Empty;
